requirements.txt
Django==1.6.2
psycopg2==2.5.2
Pillow==2.3.0
django-ratelimit==0.5.0